pyproject.toml
[tool.poetry]
name = "asgi-ml"
version = "0.1.0"
description = ""
authors = ["Vitaly Petrov <VitaliyPetrov18@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92"
aiocache = "0.12.0"
uvicorn = "^0.18.2"
ujson = "^5.4.0"
python-dotenv = "^0.1"
pytest = "^7.2"
pytest-asyncio = "^0.20.3"
pytest-pretty = "^1.1.0"
starlette-prometheus = "^0.9.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
black = "^22.3.0"
isort = "^5.9.3"

[tool.black]
line-length = 80


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
